fix(wpt): prevent snapshot node handle reuse
Reissue handles for body descendants after innerHTML re-layout, preserve the body and unaffected external nodes, and reject stale wrappers. Add a regression test for reused source NodeIds and JS wrapper identity. Remove a useless float conversion found by Clippy.

Beads: raikiri-spike-4nhl.10

+++ crates/raikiri-wpt/src/reftest.rs
@@ -1991,7 +1991,7 @@ fn render_raikiri_pages_inner_with_snapshot(
                     let width = f64::from(width);
                     let height = f64::from(height);
                     Some(ElementGeometry {
-                        offset_height: f64::from(height.round()),
+                        offset_height: height.round(),
                         bounding_client_rect: raikiri_js::DomRect {
                             left,
                             top,
